removed hls streams extractor
In resolved video processing was detection, if stream is hls. If yes, then engine tries to download master playlist and extracts streams from it.
This can't be generalised. Master playlist can have separate audio and subtitle tracks and simply extracting streams will break playback.
This should be a task to addon - only addon can decide if it is safe operation or not.